requirements.txt
asgiref==3.5.0
attr==0.3.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.1.0
platformdirs==2.5.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
